Add pure entry to manually test in jest
