+++ __tests__/pure.js
@@ -1,0 +1,5 @@
+require('../pure');
+
+test('it should be pure', () => {
+  expect(jest.isMockFunction(console.log)).toBe(false);
+});
